+++ src/WireMock.Net/Matchers/LinqMatcher.cs
@@ -9,8 +9,9 @@ namespace WireMock.Matchers
     /// <summary>
     /// System.Linq.Dynamic.Core Expression Matcher
     /// </summary>
+    /// <inheritdoc cref="IObjectMatcher"/>
     /// <inheritdoc cref="IStringMatcher"/>
-    public class LinqMatcher : IStringMatcher
+    public class LinqMatcher : IObjectMatcher, IStringMatcher
     {
         private readonly string[] _patterns;
 
@@ -117,7 +118,6 @@ namespace WireMock.Matchers
             }
 
             return MatchBehaviourHelper.Convert(MatchBehaviour, match);
-
         }
 
         /// <inheritdoc cref="IStringMatcher.GetPatterns"/>