+++ test/WireMock.Net.Tests/Serialization/MatcherModelMapperTests.cs
@@ -64,7 +64,7 @@ namespace WireMock.Net.Tests.Serialization
         }
 
         [Fact]
-        public void MatcherModelMapper_Map_ExactObjectMatcher_Pattern()
+        public void MatcherModelMapper_Map_ExactObjectMatcher_ValidBase64StringPattern()
         {
             // Assign
             var model = new MatcherModel
@@ -78,6 +78,20 @@ namespace WireMock.Net.Tests.Serialization
 
             // Assert
             Check.That(matcher.ValueAsBytes).ContainsExactly(new byte[] { 115, 116, 101, 102 });
+        }
+
+        [Fact]
+        public void MatcherModelMapper_Map_ExactObjectMatcher_InvalidBase64StringPattern()
+        {
+            // Assign
+            var model = new MatcherModel
+            {
+                Name = "ExactObjectMatcher",
+                Patterns = new object[] { "_" }
+            };
+
+            // Act & Assert
+            Check.ThatCode(() => _sut.Map(model)).Throws<ArgumentException>();
         }
 
         [Fact]